Remove AddProduct and ProductDetails components; update routes and sidebar paths
