--- src/pages/products/AddProductPage.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import AddProduct from '../../components/products/AddProduct';
-
-const AddProductPage = () => {
-  return <AddProduct />;
-};
-
-export default AddProductPage;
